Adds options,controls,and only one point per pulse
Adds options when starting the plugin for Analog input, and trigger pins.

Adds controls to change the sample rate (independent of averaging), and to
start/stop the experiment.
stopping the experiment resets all events and deregisters from all
associated channels

Only one data point is displayed per pulse. If more than one data samples
are sent during a pulse, they are averaged together into a single data
point.

The plugin *should* calculate the needed on-device average and use that to
send only a single sample to the PC, however this seems to work only some
of the time.